rename executor to scheduler
